refactor type OIDMatcher to utils package
